requirements.txt
# Bibliothèques de base pour la manipulation de données
pandas==2.0.3
numpy==1.24.3
scipy==1.10.1

# Extraction de données financières
yfinance==0.2.28
alpha_vantage==2.3.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
xlrd==2.0.1

# Traitement distribué avec Spark
pyspark==3.4.1
pyarrow==12.0.1  # Pour l'intégration efficace entre Pandas et Spark
findspark==2.0.1  # Pour localiser l'installation Spark

# Transformation et analyse de données
scikit-learn==1.3.0
statsmodels==0.14.0
ta==0.10.2  # Technical Analysis library
pandas-ta==0.3.14b0  # Alternative Technical Analysis library
pykalman==0.9.5  # Kalman filtering

# Visualisation
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.1
pandas>=1.3.0

# Base de données
psycopg2-binary==2.9.6  # PostgreSQL adapter
sqlalchemy==2.0.19
alembic==1.11.1  # Migrations de base de données
pyyaml>=6.0

# Orchestration
apache-airflow==2.6.3
apache-airflow-providers-postgres==5.6.0
apache-airflow-providers-http==4.4.2
apache-airflow-providers-slack==7.3.1
apache-airflow-providers-email==1.1.0
apache-airflow-providers-apache-spark==4.1.1  # Pour intégrer Spark avec Airflow

# Intégration avec Tableau/Power BI
tableauserverclient==0.24
pyodbc==4.0.39  # Pour Power BI

# Génération de rapports
reportlab==4.0.4
pdfkit==1.0.0
jinja2==3.1.2

# Outils de développement
pytest==7.4.0
black==23.7.0
flake8==6.0.0
isort==5.12.0
mypy==1.4.1

# Journalisation et configuration
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.0

# Sécurité
cryptography==41.0.3
python-jose==3.3.0